+++ src/include/tileset.h
@@ -44,7 +44,7 @@ public:
   friend std::istream& operator>>(std::istream& in, TilesetTile& tilesettile);
 
 private:
-  uint id;                          // The id representing the image
+  uint16 id;                          // The id representing the image
   std::string name;                 // The display name of the image
   std::string filePath;             // The path to the image
   Glib::RefPtr<Gdk::Pixbuf> pixbuf; // A pixbuf created from the image
@@ -58,7 +58,7 @@ class Tileset
 public:
   std::string name{""};
   std::string filepath{""};
-  std::unordered_map<uint, TilesetTile> tile;
+  std::unordered_map<uint16, TilesetTile> tile;
 
   Tileset(const std::string fpath = "");
